tourist_archive/requirements.txt
asgiref==3.5.0
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gpx-converter==2.0.0
gpxpy==1.5.0
numpy==1.22.3
pandas==1.4.2
psycopg2==2.9.3
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
